# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "FriendlyNote CRUD",
        "description": "My User Project Application API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:8000",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/enclosure/share": {
            "post": {
                "tags": [
                    "Enclosure"
                ],
                "parameters": [
                    {
                        "description": "User email and enclosure ID to share",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Enclosure"
                                }
                            }
                        },
                        "required": true
                    }
                ],
                "summary": "Share an enclosure with a user",
                "description": "Share an existing enclosure with another user by specifying their email and the ID of the enclosure",
                "operationId": "shareEnclosure",
                "responses": {
                    "200": {
                        "description": "Enclosure shared successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Enclosure shared"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body or parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bad request"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User or enclosure not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User or enclosure not found"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Note": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "body": {
                        "type": "string"
                    },
                    "author": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "title",
                    "body",
                    "author"
                ]
            },
...
            "Enclosure": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "The name of the enclosure"
                    },
                    "notes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Note"
                        },
                        "description": "The notes associated with the enclosure"
                    },
                    "creator": {
                        "type": "string",
                        "description": "The ID of the user who created the enclosure"
                    },
                    "dateCreated": {
                        "type": "string",
                        "format": "date-time",
                        "description": "The date and time when the enclosure was created"
                    }
                },
                "required": [
                    "name",
                    "notes",
                    "creator",
                    "dateCreated"
                ]
            },
...
        }
    }
}